+++ src/local.hpp
@@ -407,6 +407,7 @@ void DetectOverlappingLines(void);
 void DetectPolyobjSectors(buildinfo_t &config);
 
 // pruning routines
+void ClearNewVertices(void);
 void PruneVerticesAtEnd(void);
 
 // computes the wall tips for all of the vertices
@@ -517,7 +518,7 @@ quadtree_c *TreeFromSegList(seg_t *list);
 void BuildNodes(seg_t *list, int depth, bbox_t *bounds, node_t **N, subsec_t **S, double split_cost, bool fast, bool analysis);
 
 // compute the height of the bsp tree, starting at 'node'.
-int ComputeBspHeight(const node_t *node);
+size_t ComputeBspHeight(const node_t *node);
 
 // put all the segs in each subsector into clockwise order, and renumber
 // the seg indices.