+++ src/main.cpp
@@ -24,8 +24,6 @@
 #include "core.hpp"
 
 #include <cstring>
-#include <format>
-#include <fstream>
 #include <string>
 #include <vector>
 
@@ -48,70 +46,8 @@ struct map_range_t
 };
 
 static std::vector<map_range_t> map_list;
-static std::vector<std::string> analysis_csv;
 
 buildinfo_t config;
-
-//------------------------------------------------------------------------
-void AnalysisSetupFile(const char *filepath)
-{
-  auto csv_path = std::string(filepath);
-
-  auto ext_pos = FindExtension(filepath);
-  if (ext_pos > 0)
-  {
-    csv_path.resize(ext_pos);
-  }
-
-  csv_path += ".csv";
-  FileClear(csv_path.c_str());
-
-  std::string line = "map_name,is_fast,split_cost,num_segs,num_subsecs,num_nodes,size_left,size_right";
-  analysis_csv.push_back(line);
-}
-
-void AnalysisPushLine(size_t level_index, bool is_fast, double split_cost, size_t segs, size_t subsecs, size_t nodes,
-                      int32_t left_size, int32_t right_size)
-{
-  std::string line = std::format("{},{},{},{},{},{},{},{}", GetLevelName(level_index), is_fast, split_cost, segs, subsecs,
-                                 nodes, left_size, right_size);
-  analysis_csv.push_back(line);
-}
-
-// writes out for current file
-// expects AnalysisPushLine to have been called with all 0-32 split costs during node-building
-void WriteAnalysis(const char *filename)
-{
-  auto csv_path = std::string(filename);
-
-  if (const auto ext_pos = FindExtension(filename); ext_pos > 0)
-  {
-    csv_path.resize(ext_pos);
-  }
-
-  csv_path += ".csv";
-
-  // Append to fresh CSV
-  auto csv_file = std::ofstream(csv_path.c_str(), std::ios::app);
-
-  if (!csv_file.is_open())
-  {
-    PrintLine(LOG_WARN, "[%s] Couldn't open file %s for writing.", __func__, csv_path.c_str());
-    return;
-  }
-
-  for (const auto &line : analysis_csv)
-  {
-    csv_file << line << '\n';
-  }
-
-  csv_file.close();
-
-  // Flush out from memory
-  analysis_csv.clear();
-
-  PrintLine(LOG_NORMAL, "[%s] Successfully finished writing data to CSV file %s.", __func__, csv_path.c_str());
-}
 
 //------------------------------------------------------------------------
 
@@ -291,7 +227,7 @@ void VisitFile(const char *filename)
 
   if (config.analysis)
   {
-    AnalysisSetupFile(filename);
+    SetupAnalysisFile(filename);
   }
 
   PrintLine(LOG_NORMAL, "Building %s", filename);