+++ src/node.cpp
@@ -1615,19 +1615,6 @@ subsec_t *CreateSubsec(quadtree_c *tree)
   return sub;
 }
 
-int ComputeBspHeight(const node_t *node)
-{
-  if (node == nullptr)
-  {
-    return 1;
-  }
-
-  int right = ComputeBspHeight(node->r.node);
-  int left = ComputeBspHeight(node->l.node);
-
-  return std::max(left, right) + 1;
-}
-
 void BuildNodes(seg_t *list, int depth, bbox_t *bounds, node_t **N, subsec_t **S, double split_cost, bool fast, bool analysis)
 {
   *N = nullptr;